# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for JuanitaMusic discord bot",
        "title": "JuanitaAPI",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/guild": {
            "post": {
                "tags": [
                    "Guild"
                ],
                "summary": "Create a new guild",
                "operationId": "CreateGuild",
                "parameters": [
                    {
                        "description": "Guild to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Guild"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Guild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/guild/{guildId}": {
            "put": {
                "tags": [
                    "Guild"
                ],
                "summary": "Update a guild",
                "operationId": "UpdateGuild",
                "parameters": [
                    {
                        "description": "Guild to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Guild"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Guild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/guild/{guildId}/language": {
            "patch": {
                "tags": [
                    "Guild"
                ],
                "summary": "Update a guild language",
                "operationId": "UpdateGuildLanguage",
                "parameters": [
                    {
                        "description": "Guild language to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GuildLanguagePatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GuildLanguagePatch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/guilds": {
            "get": {
                "tags": [
                    "Guild"
                ],
                "summary": "Get all guilds",
                "operationId": "GetGuilds",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit the amount of guilds returned",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Guild"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Alias": {
            "type": "object",
            "required": [
                "alias",
                "name",
                "playlistid"
            ],
            "properties": {
                "alias": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "playlistid": {
                    "type": "string"
                }
            }
        },
...
        "models.Guild": {
            "type": "object",
            "required": [
                "aliases",
                "id",
                "language",
                "name",
                "permissionrole"
            ],
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Alias"
                    }
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "$ref": "#/definitions/models.Language"
                },
                "name": {
                    "type": "string"
                },
                "permissionrole": {
                    "type": "string"
                }
            }
        },
        "models.GuildLanguagePatch": {
            "type": "object",
            "required": [
                "language"
            ],
            "properties": {
                "language": {
                    "$ref": "#/definitions/models.Language"
                }
            }
        },
        "models.Language": {
            "type": "string",
            "enum": [
                "unknown"
            ],
            "x-enum-varnames": [
                "Unknown"
            ]
        },
...
    }
}